Fix infinite symbolic links
We're using filter_builder in CI by installing it into `vendor/bundle`
via bundlers `--path=` option along with other gems. Sometimes loading
the app will cause in infinite loop. This is because the
`spec/dummy/spec` path is a symbolic link to `../../spec`. So ruby goes
to load `/spec`, which loads `/spec/dummy`, which loads
`/spec/dummy/spec` which loads `/spec` again.

This fixes the issue by:

- Removing the `/spec/dummy/spec` symbolic link
- Moving the fabricators to `/spec/dummy/spec/fabricators` so they can
be found by the test suite